Fix integration test failures: 159/164 passing (97.0%)
Major fixes:
- Created conftest.py with global_config, device, brain_config fixtures
- Fixed checkpoint JSON serialization for torch.device and nested objects
- Fixed striatum pathway weight linking in checkpoint restore
- Fixed LayeredCortex forward() API to use dict inputs
- Fixed STP module state initialization (reset_state vs reset)
- Fixed Striatum.evaluate_state() API (removed goal_context parameter)
- Fixed PFC type mismatches (bool to float conversions)
- Updated cortex output size expectations (LayerRatios 2:1)
- Added hippocampus.retrieve_similar() alias method
- Enhanced checkpoint serialization with recursive _serialize_value()
- Fixed full brain state construction in select_action()
- Modified coordinator to handle full concatenated state correctly
- Fixed consolidation manager to pass dict input to striatum

Improved from 84 failures (50%) to 5 failures (97.0%).

Remaining issues (5 tests):
- CA3 persistent activity not building up (ACh modulation)
- Checkpoint tensor serialization (deeply nested objects)
- D1 delay buffer preservation in checkpoints

+++ src/thalia/regions/prefrontal/prefrontal.py
@@ -1239,7 +1239,8 @@ class Prefrontal(NeuralRegion):
                 prediction = state_projection * action_modulation
             else:
                 # Fallback: simple recurrent prediction
-                prediction = self.rec_weights @ current_state
+                # Convert bool spikes to float for matrix multiplication
+                prediction = self.rec_weights @ current_state.float()
 
         # Apply nonlinearity (tanh to keep bounded)
         prediction = torch.tanh(prediction)
